+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * `pre_cell` and `post_cell` keywords in table metadata to apply markdown to entire table
 * `pylingdocs author-config` command
+* `add_bib` option for adding external sources
 
 ### Removed
 * text cldf metadata (now in [clld-corpus-plugin](https://github.com/fmatter/clld-corpus-plugin/blob/main/src/clld_corpus_plugin/cldf/TextTable-metadata.json))